refactored file directory
