+++ lib/src/model/business_model.dart
